got snow working, centered CSS, added a file to make it easier to read
